# api/docs/swagger.yaml
definitions:
...
  models.CreateStorage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      product_id:
        type: string
    type: object
  models.CreateStorageTransaction:
    properties:
      created_at:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: number
      staff_id:
        type: string
      storage_transaction_type:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      status_code:
        type: integer
    type: object
...
  models.Storage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
    type: object
  models.StorageTransaction:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: number
      staff_id:
        type: string
      storage_transaction_type:
        type: string
      updated_at:
        type: string
    type: object
  models.StorageTransactionsResponse:
    properties:
      count:
        type: integer
      storage_transactions:
        items:
          $ref: '#/definitions/models.StorageTransaction'
        type: array
    type: object
  models.StoragesResponse:
    properties:
      count:
        type: integer
      repositories:
        items:
          $ref: '#/definitions/models.Storage'
        type: array
    type: object
...
  models.UpdateStorage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      id:
        type: string
      product_id:
        type: string
    type: object
  models.UpdateStorageTransaction:
    properties:
      id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: number
      staff_id:
        type: string
      storage_transaction_type:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /storage:
    post:
      consumes:
      - application/json
      description: create a new storage
      parameters:
      - description: storage
        in: body
        name: storage
        schema:
          $ref: '#/definitions/models.CreateStorage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new storage
      tags:
      - storage
  /storage/{id}:
    delete:
      consumes:
      - application/json
      description: delete storage
      parameters:
      - description: storage_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete storage
      tags:
      - storage
    get:
      consumes:
      - application/json
      description: get storage by id
      parameters:
      - description: storage_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get storage by id
      tags:
      - storage
    put:
      consumes:
      - application/json
      description: update storage
      parameters:
      - description: storage_id
        in: path
        name: id
        required: true
        type: string
      - description: storage
        in: body
        name: storage
        schema:
          $ref: '#/definitions/models.UpdateStorage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update storage
      tags:
      - storage
  /storage_transaction:
    post:
      consumes:
      - application/json
      description: create a new  storage_transaction
      parameters:
      - description: storage_transaction
        in: body
        name: storage_transaction
        schema:
          $ref: '#/definitions/models.CreateStorageTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StorageTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new storage_transaction
      tags:
      - storage_transaction
  /storage_transaction/{id}:
    delete:
      consumes:
      - application/json
      description: delete storage_transaction
      parameters:
      - description: storage_transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete storage_transaction
      tags:
      - storage_transaction
    get:
      consumes:
      - application/json
      description: get storage_transaction by id
      parameters:
      - description: storage_transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StorageTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get storage_transaction by id
      tags:
      - storage_transaction
    put:
      consumes:
      - application/json
      description: update storage_transaction
      parameters:
      - description: storage_transaction_id
        in: path
        name: id
        required: true
        type: string
      - description: storage_transactions
        in: body
        name: storage_transactions
        schema:
          $ref: '#/definitions/models.UpdateStorageTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StorageTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update storage_transaction
      tags:
      - storage_transaction
  /storage_transactions:
    get:
      consumes:
      - application/json
      description: get storage_transaction list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StorageTransactionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get storage_transaction list
      tags:
      - storage_transaction
  /storages:
    get:
      consumes:
      - application/json
      description: get storage list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StoragesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get storage list
      tags:
      - storage
...
swagger: "2.0"